Remove warnings

+++ core/src/main/scala/io/github/saucam/shiva/common/DistanceCalculatorImpl.scala
@@ -1,9 +1,9 @@
 package io.github.saucam.shiva.common
 
+import scala.math.pow
+
 import breeze.linalg._
 import breeze.numerics.abs
-
-import scala.math.pow
 
 class InnerProductFloat extends DistanceCalculator[Float] {
   override def computeDistance(u: Vector[Float], v: Vector[Float]): Float = 1 - (u dot v)
@@ -71,7 +71,7 @@ class MinkowskiDistanceFloat(p: Int) extends DistanceCalculator[Float] {
   override def computeDistance(u: Vector[Float], v: Vector[Float]): Float = {
     val diff = u - v
     val absDiff = abs(diff)
-    val powDiff = absDiff.mapValues(d => pow(d, p).toFloat)
+    val powDiff = absDiff.mapValues(d => pow(d.toDouble, p.toDouble))
     val sumDiff = sum(powDiff)
     pow(sumDiff, 1.0 / p).toFloat
   }
@@ -81,7 +81,7 @@ class MinkowskiDistanceDouble(p: Int) extends DistanceCalculator[Double] {
   override def computeDistance(u: Vector[Double], v: Vector[Double]): Double = {
     val diff = u - v
     val absDiff = abs(diff)
-    val powDiff = absDiff.mapValues(d => pow(d, p))
+    val powDiff = absDiff.mapValues(d => pow(d, p.toDouble))
     val sumDiff = sum(powDiff)
     pow(sumDiff, 1.0 / p)
   }
